feat: IA-1609 - implement request rate-limiting to prevent denial-of-service attacks
This addresses a CodeQL issue which flagged that the auth routes were not covered by rate-limiting.
A global limit on all routes was chosen as reasonable use of the app functionality shouldn't require many server calls
(once logged in data is fetched only at the user's request). A limit of 20 requests per IP seemed suitable in order
to not interfere with genuine use of the tool while providing some protection against denial-of-service.